[NFC] compositing: extract types into new compositing_traits crate
