+++ components/constellation/pipeline.rs
@@ -7,9 +7,9 @@ use crate::sandboxing::{spawn_multiprocess, UnprivilegedContent};
 use background_hang_monitor::HangMonitorRegister;
 use bluetooth_traits::BluetoothRequest;
 use canvas_traits::webgl::WebGLPipeline;
-use compositing::compositor_thread::Msg as CompositorMsg;
-use compositing::CompositionPipeline;
-use compositing::CompositorProxy;
+use compositing_traits::CompositionPipeline;
+use compositing_traits::CompositorProxy;
+use compositing_traits::Msg as CompositorMsg;
 use crossbeam_channel::{unbounded, Sender};
 use devtools_traits::{DevtoolsControlMsg, ScriptToDevtoolsControlMsg};
 use embedder_traits::EventLoopWaker;
