+++ components/servo/Cargo.toml
@@ -40,6 +40,7 @@ bluetooth_traits = { path = "../bluetooth_traits" }
 canvas = { path = "../canvas", default-features = false }
 canvas_traits = { path = "../canvas_traits" }
 compositing = { path = "../compositing", features = ["gl"] }
+compositing_traits = { path = "../compositing_traits", features = ["gl"] }
 constellation = { path = "../constellation" }
 crossbeam-channel = { workspace = true }
 devtools = { path = "../devtools" }
