+++ components/webdriver_server/Cargo.toml
@@ -12,7 +12,7 @@ path = "lib.rs"
 
 [dependencies]
 base64 = { workspace = true }
-compositing = { path = "../compositing" }
+compositing_traits = { path = "../compositing_traits" }
 cookie = { workspace = true }
 crossbeam-channel = { workspace = true }
 euclid = { workspace = true }
